requirements.txt
fastapi
uvicorn
httpx
pyjwt
pydantic-settings
